+++ CHANGELOG.md
@@ -47,7 +47,7 @@ The displayed date is in the format `DD-MM-YYYY`
 - feat: New cmds `chanid`, `roleid`, `serverid`, `userid`
 - feat: New cmd `autorole` & new plugin autorole
 - feat: New cmd `reverse`
-- feat: New slash cmd `ping`
+- feat: New slash commands `ping`, `staff`
 - feat: New cmd `config`
 
 ### Changes